Issue #7: add folder for real-time simulation.
- Update and adapt the classes State and StateMachine to run in real-time simulations;
- Real time simulation folder is test_state_machines;
- Update file in test folder updates the model in src and test_state_machine folders.

+++ src/model/State.m
@@ -1,6 +1,6 @@
 classdef State 
     properties
-        number double
+        number int32
         name char
         activeEvents logical
     end
